Вынес провайдер коллекции для второго экрана формы
Исправил косяк на 2 этапе с невыбранным праздником
