+++ holidayNow/Scenes/SecondFormScreen/SecondFormViewController.swift
@@ -5,7 +5,8 @@ final class SecondFormViewController: UIViewController {
     // MARK: - Dependencies:
     weak var coordinator: CoordinatorProtocol?
     
-    private var viewModel: SecondFormViewModelProtocol
+    private let viewModel: SecondFormViewModelProtocol
+    private let collectionProvider: SecondFormCollectionProvider
     
     // MARK: - Constants and Variables:
     private enum SecondFormUIConstants {
@@ -25,7 +26,7 @@ final class SecondFormViewController: UIViewController {
         return label
     }()
     
-    private lazy var secondFormCollectionView: BaseCollectionView = {
+    private(set) lazy var secondFormCollectionView: BaseCollectionView = {
         let collection = BaseCollectionView()
         collection.register(BaseCollectionViewCell.self,
                             forCellWithReuseIdentifier: Resources.Identifiers.formInterestCollectionVewCell)
@@ -35,9 +36,6 @@ final class SecondFormViewController: UIViewController {
                             forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader,
                             withReuseIdentifier: Resources.Identifiers.collectionReusableView)
         collection.isScrollEnabled = false
-        collection.dataSource = self
-        collection.delegate = self
-        
         collection.backgroundColor = .whiteDay
         
         return collection
@@ -49,9 +47,10 @@ final class SecondFormViewController: UIViewController {
     private lazy var screenScrollView = UIScrollView()
     
     // MARK: - Lifecycle:
-    init(coordinator: CoordinatorProtocol?, viewModel: SecondFormViewModelProtocol) {
+    init(coordinator: CoordinatorProtocol?, viewModel: SecondFormViewModelProtocol, collectionProvider: SecondFormCollectionProvider) {
         self.coordinator = coordinator
         self.viewModel = viewModel
+        self.collectionProvider = collectionProvider
         super.init(nibName: nil, bundle: nil)
     }
     
@@ -64,6 +63,7 @@ final class SecondFormViewController: UIViewController {
         setupViews()
         setupConstraints()
         setupTargets()
+        setupCollectionProvider()
         setupObservers()
         
         bind()
@@ -96,7 +96,7 @@ final class SecondFormViewController: UIViewController {
             if self.secondFormCollectionView.indexPathsForSelectedItems?.count == 0 {
                 self.resumeOnMainThread(self.continueButton.block, with: ())
             } else {
-                self.resumeOnMainThread(self.continueButton.unblock, with: ())
+                self.resumeOnMainThread(self.checkSelectedCell, with: ())
             }
         }
     }
@@ -112,34 +112,18 @@ final class SecondFormViewController: UIViewController {
         secondFormCollectionView.selectItem(at: indexPath, animated: true, scrollPosition: .top)
     }
     
-    private func getParticularCell(isDefault: Bool, indexPath: IndexPath, isHoliday: Bool) -> UICollectionViewCell {
-        if isDefault {
-            let defaultCell: BaseCollectionViewCell = secondFormCollectionView.dequeueReusableCell(
-                indexPath: indexPath,
-                with: Resources.Identifiers.formInterestCollectionVewCell)
-            defaultCell.delegate = self
-
-            if isHoliday {
-                let model = viewModel.holidaysObserver.wrappedValue.holidays[indexPath.row]
-                defaultCell.setupInterestModel(model: CellModel(name: model.name, image: model.image))
-                
-                return defaultCell
-            } else {
-                let model = viewModel.intonations.intonations[indexPath.row]
-                defaultCell.setupInterestModel(model: CellModel(name: model.name, image: model.image))
-                
-                return defaultCell
-            }
+    private func checkSelectedCell() {
+        guard let selectedIndexes = secondFormCollectionView.indexPathsForSelectedItems else { return }
+        if selectedIndexes.contains(where: { $0.section == 0 }) {
+            continueButton.unblock()
         } else {
-            let defaultEnterCell: BaseCollectionViewEnterCell = secondFormCollectionView.dequeueReusableCell(
-                indexPath: indexPath,
-                with: Resources.Identifiers.formEnterInterestCollectionVewCell)
-            
-            defaultEnterCell.delegate = self
-            defaultEnterCell.setupCellWidht(value: view.frame.width)
-            
-            return defaultEnterCell
+            continueButton.block()
         }
+    }
+    
+    private func setupCollectionProvider() {
+        secondFormCollectionView.dataSource = collectionProvider
+        secondFormCollectionView.delegate = collectionProvider
     }
     
     // MARK: - Objc Methods:
@@ -187,112 +171,6 @@ extension SecondFormViewController: BaseCollectionViewEnterCellDelegate {
             controlStateWarningLabel(label: warningLabel,
                                      isShow: false)
         }
-    }
-}
-
-// MARK: - UICollectionViewDataSource:
-extension SecondFormViewController: UICollectionViewDataSource {
-    func numberOfSections(in collectionView: UICollectionView) -> Int {
-        2
-    }
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        switch section {
-        case 0:
-            return viewModel.holidaysObserver.wrappedValue.holidays.count + 1
-        case 1:
-            return viewModel.intonations.intonations.count
-        default:
-            return  0
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        switch indexPath.section {
-        case 0:
-            if indexPath.row == viewModel.holidaysObserver.wrappedValue.holidays.count {
-                guard let cell = getParticularCell(
-                    isDefault: false,
-                    indexPath: indexPath,
-                    isHoliday: true) as? BaseCollectionViewEnterCell else { return UICollectionViewCell() }
-                
-                return cell
-            } else {
-                guard let cell = getParticularCell(
-                    isDefault: true,
-                    indexPath: indexPath,
-                    isHoliday: true) as? BaseCollectionViewCell else { return UICollectionViewCell() }
-                
-                return cell
-            }
-        case 1:
-            guard let cell = getParticularCell(
-                isDefault: true,
-                indexPath: indexPath,
-                isHoliday: false) as? BaseCollectionViewCell else { return UICollectionViewCell() }
-            
-            return cell
-        default:
-            return UICollectionViewCell()
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
-        let indexPath = IndexPath(row: 0, section: section)
-        let headerView = self.collectionView(collectionView,
-                                             viewForSupplementaryElementOfKind: UICollectionView.elementKindSectionHeader,
-                                             at: indexPath)
-
-        return headerView.systemLayoutSizeFitting(CGSize(width: collectionView.frame.width,
-                                                         height: UIView.layoutFittingExpandedSize.height),
-                                                  withHorizontalFittingPriority: .required,
-                                                  verticalFittingPriority: .fittingSizeLevel)
-    }
-}
-
-// MARK: - UICollectionViewDelegateFlowLayout:
-extension SecondFormViewController: UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        CGSize(width: collectionView.frame.width, height: UIConstants.inputHeight)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        guard let selectedIndex = collectionView.indexPathsForSelectedItems else { return }
-        
-        selectedIndex.forEach { index in
-            if index.section == indexPath.section && index.row != indexPath.row {
-                collectionView.deselectItem(at: index, animated: true)
-            }
-        }
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        var text: String
-        var id: String
-        
-        switch indexPath.section {
-        case 0:
-            text = viewModel.holidaysObserver.wrappedValue.name
-        case 1:
-            text = viewModel.intonations.name
-        default:
-            text = ""
-        }
-        
-        switch kind {
-        case UICollectionView.elementKindSectionHeader:
-            id = Resources.Identifiers.collectionReusableView
-        default:
-            id = ""
-        }
-        
-        guard let headerView = collectionView.dequeueReusableSupplementaryView(
-            ofKind: kind,
-            withReuseIdentifier: id,
-            for: indexPath) as? BaseCollectionViewReusableView else { return UICollectionReusableView() }
-        
-        headerView.setupLabelName(with: text)
-        
-        return headerView
     }
 }
